+++ src/lib/controls/list-item/listItemComponent.angular.spec.ts
@@ -11,11 +11,11 @@ import { NotificationType } from '../../models/notificationType.model';
 
 import {
   ListItemComponent, ListItemHeaderComponent, ListColumnComponent,
-  ListItemContentComponent, ListItemJqeuryHelper
+  ListItemContentComponent
 } from '../../index';
 
 @Component({
-  selector: 'test',
+  selector: 'vgr-test',
   template: `
           <vgr-list-item>
             <vgr-list-item-header>
@@ -424,6 +424,7 @@ describe('ListItemComponent', () => {
       beforeEach(() => {
         spyOn(component.expandedChanged, 'emit');
         component.notification = { message: 'Nu sparar vi', icon: 'vgr-icon-ok-check ', type: NotificationType.ShowOnCollapse, removeWhenDone: true } as RowNotification;
+        listItemComponentFixture.detectChanges();
       });
       it('notification is displayed', () => {
         expect(component.notificationVisible).toBe(true);
@@ -435,7 +436,8 @@ describe('ListItemComponent', () => {
 
       describe('and notification is collapsing after 1,4s', () => {
         beforeEach(() => {
-          jasmine.clock().tick(1400);
+          jasmine.clock().tick(3401);
+
         });
         it('content is collapsed', () => {
           expect(component.collapsed).toBe(true);